backend/requirements.txt
Flask==2.2.5
web3==6.9.0
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
